#70 checked the result of cielab.solve_chroma.
